Adopt ruff 0.16 and modernize the codebase to its default rules
ruff 0.16 broadens the default rule set (pyupgrade, isort, comprehensions and
more). Bump the dev pin to ruff>=0.16.0 and apply the resulting fixes across
the tree: PEP 585 built-in generic annotations, unquoted annotations under
from-future annotations, sorted imports and the assorted small lint fixes. A
re-export shim that relied on redundant import aliases now declares __all__
explicitly. The api reference is regenerated for the modernized signatures.
No runtime behavior changes; the full test suite passes.

+++ tests/metrology/test_time_frequency.py
@@ -27,7 +27,7 @@ REF_PRESSURE = 2e-5
 def _tone(
     fs: float, duration: float, freq: float, amplitude: float
 ) -> np.ndarray:
-    t = np.arange(int(round(fs * duration))) / fs
+    t = np.arange(round(fs * duration)) / fs
     return amplitude * np.cos(2.0 * np.pi * freq * t)
